backend/requirements.txt
annotated-doc==0.0.2
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
black==25.9.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.120.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.11.1
mypy_extensions==1.1.0
openai==2.6.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.2.0
PyYAML==6.0.3
redis==7.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
